pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphasearch"
version = "0.3.0"
description = "AlphaGo-inspired MCTS for document retrieval. No vectors, no embeddings, just reasoning."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "KnightOwl"},
]
keywords = ["mcts", "document-retrieval", "rag", "pdf", "search", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "openai>=1.30.0",
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/CyberGhost007/alphasearch"
Repository = "https://github.com/CyberGhost007/alphasearch"
Issues = "https://github.com/CyberGhost007/alphasearch/issues"

[project.scripts]
alphasearch = "main:main"

[tool.setuptools.packages.find]
include = ["treerag*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
